% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flashPoint}
\alias{flashPoint}
\title{Create a location to flash messages}
\usage{
flashPoint(id = "")
}
\arguments{
\item{id}{Optional id of div}
}
\description{
Called in the UI to create a spot for flashing alerts.
}

