% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flash.R
\name{flash}
\alias{flash}
\title{Flash a message}
\usage{
flash(session, text, type = "info")
}
\arguments{
\item{session}{session}

\item{text}{text}

\item{type}{type: boostrap types ['info', 'warning', 'alert', 'success']}
}
\description{
Flash an alert message bootstrap style
Will display the alert in first of the following containers by order
of preference: 'flash-container', 'container', 'container-fluid', 'body[0]'.
Use the flash-container though, since it will will be styled to come to the top.
}

